modularize qa-tools
