+++ pallets/qa-tools/src/pallets/mod.rs
@@ -1,0 +1,4 @@
+//! Pallet-separated tools. To avoid naming clashes and
+//! to distinguish logic related to different pallets.
+
+pub mod order_book_tools;
